dra scheduler plugin: move extended resources functions into separate file
